+++ Realm/Tests/RLMTestObjects.h
@@ -85,7 +85,7 @@ RLM_ARRAY_TYPE(IntObject)
 @property NSData       *binaryCol;
 @property NSDate       *dateCol;
 @property bool          cBoolCol;
-@property long long     longCol;
+@property int64_t     longCol;
 @property id            mixedCol;
 @property StringObject *objectCol;
